Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2pydantic==2.7.1
pydantic-settings==2.2.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
prisma==0.10.0
python-dotenv==1.0.0
asyncpg==0.30.0
psycopg2-binary==2.9.9
pydantic==2.9.9

